refactor(ddl): decouple PostgreSQL renderer
