+++ internal/ddl/postgres/renderer_fixes_test.go
